ref(history.currency-chart): rename to kebab case, remove unnessary scope, fix twice chart rendering
